[libpsyseg] Add proof of concept library
Basically something I had in mind since forever, and never sat down to
bash out.  It's a very simple idea of prepending files with certain
metadata so that you can have a portable way to sort them.

The files I'm interested in are usually media files such as
pictures. Pictures have many formats, and over time have evolved to
allow for the possibility of metadata.

However, different formats have their own ways of implementing
metadata, and sometimes said formats will not support metadata at all,
which make maintaining said collections much more annoying.

So I wondered how easy it would be, to write a container like format
for image files, that could store the metadata in the header, and thus
make it portable quite easilly instead of fiddling around with a
database (which you'd have to remember carrying over to another
computer for example).

The next question is how to view these files, as they'll most likely
break the common viewers. For this I have two ideas:
- patch a custom `feh` to simply skip over the metadata, and use its
  usual viewing routines
- write a simple SDL viewer for media

Another aspect that I want to piggyback is appending gpg signatures to
pictures. This is for another idea I had, where artists could claim
certain works as their own (as there is quite a bit of art theft on
the internet).

That's it for now I guess.

Signed-off-by: Simon Symeonidis <lethaljellybean@gmail.com>